# apps/docs/content/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/customers": {
      "post": {
        "operationId": "CustomerController_create",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "example": "-state",
              "type": "string"
            }
          },
          {
            "name": "expand",
            "required": false,
            "in": "query",
            "description": "List of relations to be populated.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "select",
            "required": false,
            "in": "query",
            "description": "List of fields to select.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "filter",
            "required": false,
            "in": "query",
            "description": "Filter options",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "or",
            "required": false,
            "in": "query",
            "description": "Or options",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "nor",
            "required": false,
            "in": "query",
            "description": "Nor options",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "and",
            "required": false,
            "in": "query",
            "description": "And options",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "before",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "after",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search query",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "projection",
            "required": false,
            "in": "query",
            "description": "Filter Projection",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "countTotal",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "dryRun",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create Customer",
        "tags": [
          "customers"
        ]
      },
      "get": {
        "operationId": "CustomerController_getCustomer",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "example": "-state",
              "type": "string"
            }
          },
          {
            "name": "expand",
            "required": false,
            "in": "query",
            "description": "List of relations to be populated.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "select",
            "required": false,
            "in": "query",
            "description": "List of fields to select.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "filter",
            "required": false,
            "in": "query",
            "description": "Filter options",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "or",
            "required": false,
            "in": "query",
            "description": "Or options",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "nor",
            "required": false,
            "in": "query",
            "description": "Nor options",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "and",
            "required": false,
            "in": "query",
            "description": "And options",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "before",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "after",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search query",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "projection",
            "required": false,
            "in": "query",
            "description": "Filter Projection",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "countTotal",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "dryRun",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get Customer",
        "tags": [
          "customers"
        ]
      }
    },
    "/customers/{id}": {
      "get": {
        "operationId": "CustomerController_getOne",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "example": "-state",
              "type": "string"
            }
          },
          {
            "name": "expand",
            "required": false,
            "in": "query",
            "description": "List of relations to be populated.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "select",
            "required": false,
            "in": "query",
            "description": "List of fields to select.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "filter",
            "required": false,
            "in": "query",
            "description": "Filter options",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "or",
            "required": false,
            "in": "query",
            "description": "Or options",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "nor",
            "required": false,
            "in": "query",
            "description": "Nor options",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "and",
            "required": false,
            "in": "query",
            "description": "And options",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "before",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "after",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search query",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "projection",
            "required": false,
            "in": "query",
            "description": "Filter Projection",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "countTotal",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "dryRun",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get Customer",
        "tags": [
          "customers"
        ]
      },
      "put": {
        "operationId": "CustomerController_updateCustomer",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "example": "-state",
              "type": "string"
            }
          },
          {
            "name": "expand",
            "required": false,
            "in": "query",
            "description": "List of relations to be populated.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "select",
            "required": false,
            "in": "query",
            "description": "List of fields to select.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "filter",
            "required": false,
            "in": "query",
            "description": "Filter options",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "or",
            "required": false,
            "in": "query",
            "description": "Or options",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "nor",
            "required": false,
            "in": "query",
            "description": "Nor options",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "and",
            "required": false,
            "in": "query",
            "description": "And options",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "before",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "after",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search query",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "projection",
            "required": false,
            "in": "query",
            "description": "Filter Projection",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "countTotal",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "dryRun",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchCustomerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update Customer",
        "tags": [
          "customers"
        ]
      }
    },
...
  },
  "info": {
    "title": "Propeller APIs",
    "description": "Financial Infrastructure for the China–Africa corridor",
    "version": "2024-05-01",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "CustomerIndividualInformationDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "gender": {
            "type": "string",
            "description": "Gender",
            "example": "M",
            "enum": [
              "M",
              "F"
            ]
          },
          "nationalityCode": {
            "type": "string",
            "description": "Nationality in ISO Country Code",
            "example": "NG"
          },
          "firstName": {
            "type": "string",
            "description": "First Name"
          },
          "middleName": {
            "type": "string",
            "description": "Middle Name"
          },
          "lastName": {
            "type": "string",
            "description": "Last Name"
          },
          "phoneNumber": {
            "type": "string",
            "description": "Phone Number with Calling Code",
            "example": "+2348123456789"
          },
          "dateOfBirth": {
            "type": "string",
            "description": "Date of Birth"
          },
          "email": {
            "type": "string",
            "description": "Email Address",
            "example": "customer@example.com"
          }
        },
        "required": [
          "title",
          "nationalityCode",
          "firstName",
          "lastName",
          "phoneNumber"
        ]
      },
      "CustomerIndividualAddressDto": {
        "type": "object",
        "properties": {
          "city": {
            "type": "string",
            "description": "Primary City",
            "example": "Yaba"
          },
          "state": {
            "type": "string",
            "description": "Primary Address Line One",
            "example": "Lagos"
          },
          "countryCode": {
            "type": "string",
            "description": "Primary Country Code",
            "example": "NG"
          },
          "addressLineOne": {
            "type": "string",
            "description": "Primary Address Line One",
            "example": "No 10, Adekunle Close"
          },
          "addressLineTwo": {
            "type": "string",
            "description": "Primary Address Line Two",
            "example": "Off Ciroma Rd"
          }
        },
        "required": [
          "city",
          "state",
          "countryCode",
          "addressLineOne"
        ]
      },
      "CustomerIndividualIdentityDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Identification Type",
            "example": "bvn",
            "enum": [
              "bvn",
              "drivers-license",
              "nin",
              "voters-card",
              "passport"
            ]
          },
          "id": {
            "type": "string",
            "description": "Id of the identification Document",
            "example": "A015BVP13Z"
          },
          "url": {
            "type": "string",
            "description": "Url of the uploaded Document",
            "example": "https://example.com/document.png"
          },
          "issuingCountry": {
            "type": "string",
            "description": "Issuing Country in ISO Country Code",
            "example": "NG"
          }
        },
        "required": [
          "type",
          "id",
          "issuingCountry"
        ]
      },
      "CustomerBusinessInformationDto": {
        "type": "object",
        "properties": {
          "registrationName": {
            "type": "string",
            "description": "Registration Name of the Company",
            "example": "Good Mill Ltd"
          },
          "phoneNumber": {
            "type": "string",
            "description": "Phone Number with Calling Code",
            "example": "+2348123456789"
          },
          "email": {
            "type": "string",
            "description": "Email of the Company",
            "example": "goodmill@gmail.com"
          }
        },
        "required": [
          "registrationName",
          "phoneNumber"
        ]
      },
      "CustomerCorporateIdentityDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Identification Type",
            "example": "cac",
            "enum": [
              "cac"
            ]
          },
          "id": {
            "type": "string",
            "description": "Registration Number of the Company",
            "example": "A015BVP13Z"
          },
          "url": {
            "type": "string",
            "description": "Url of the uploaded Document",
            "example": "https://example.com/document.png"
          },
          "issuingCountry": {
            "type": "string",
            "description": "Issuing Country in ISO Country Code",
            "example": "NG"
          }
        },
        "required": [
          "type",
          "id",
          "issuingCountry"
        ]
      },
      "CustomerClaimsDto": {
        "type": "object",
        "properties": {
          "individualInformation": {
            "$ref": "#/components/schemas/CustomerIndividualInformationDto"
          },
          "individualAddress": {
            "$ref": "#/components/schemas/CustomerIndividualAddressDto"
          },
          "individualIdentity": {
            "$ref": "#/components/schemas/CustomerIndividualIdentityDto"
          },
          "businessInformation": {
            "$ref": "#/components/schemas/CustomerBusinessInformationDto"
          },
          "businessIdentity": {
            "$ref": "#/components/schemas/CustomerCorporateIdentityDto"
          },
          "businessDirectors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "businessAddress": {
            "$ref": "#/components/schemas/CustomerIndividualAddressDto"
          }
        },
        "required": [
          "individualInformation",
          "individualAddress",
          "individualIdentity",
          "businessInformation",
          "businessIdentity",
          "businessDirectors",
          "businessAddress"
        ]
      },
      "CustomerDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Display Name"
          },
          "reference": {
            "type": "string",
            "description": "Customer Reference",
            "example": "customer@example.com"
          },
          "type": {
            "type": "string",
            "description": "Customer Type",
            "example": "individual",
            "enum": [
              "individual",
              "business"
            ]
          },
          "claims": {
            "$ref": "#/components/schemas/CustomerClaimsDto"
          },
          "verifications": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "metadata": {
            "type": "object"
          }
        },
        "required": [
          "name",
          "reference",
          "type",
          "claims",
          "verifications",
          "metadata"
        ]
      },
      "PatchCustomerDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Display Name"
          },
          "reference": {
            "type": "string",
            "description": "Customer Reference",
            "example": "customer@example.com"
          },
          "type": {
            "type": "string",
            "description": "Customer Type",
            "example": "individual",
            "enum": [
              "individual",
              "business"
            ]
          },
          "status": {
            "type": "string",
            "description": "Customer Status",
            "example": "active",
            "enum": [
              "inactive",
              "active"
            ]
          },
          "claims": {
            "$ref": "#/components/schemas/CustomerClaimsDto"
          },
          "verifications": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "metadata": {
            "type": "object"
          }
        }
      },
...
    }
  }
}